+++ source/Component/Graphics/2d/Sprites/Bgmap/BgmapSprite.c
@@ -115,7 +115,7 @@ void BgmapSprite::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool BgmapSprite::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool BgmapSprite::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
