+++ source/Component/Graphics/2d/Sprites/SpriteManager.c
@@ -116,7 +116,7 @@ void SpriteManager::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool SpriteManager::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool SpriteManager::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
