+++ source/Component/Graphics/3d/DirectDraw.c
@@ -1084,7 +1084,7 @@ static bool DirectDraw::isPointInsideFrustum(PixelVector point)
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool DirectDraw::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool DirectDraw::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
