+++ source/Entity/Container/Stage/Stage.c
@@ -917,7 +917,7 @@ Actor Stage::doAddChildActor(const PositionedActor* const positionedActor, bool 
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool Stage::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool Stage::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
