+++ source/Tools/SoundTest.c
@@ -280,7 +280,7 @@ void SoundTest::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool SoundTest::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool SoundTest::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
